Disable user scaling on mobile
Sadly this is taking away an accessibility feature (zooming in to see
text better).  I am doing this because that mostly didn't work right anyway.
Usually Google Maps would hijack the pinch gestures.  On the occasions
when it didn't, it could be hard to get it back to the usual scale
again after Google Maps has control of pinch gestures again.