add a quadtree
As of this commit, not yet used, and not yet working
because it's waiting for Leaflet to be committed
(nearestStopsTo uses L.LatLng's distanceTo).

It would be simple to get another implementation of the
Haversine distance formula if you want to use this
code without switching to Leaflet.

This code will become used and working in the next commit.